refactor(NodeService): move chooseTransition() and related functions to studentNodeService
